apps and category

+++ src/main/java/id/kings/stbmanager/repository/AppsRepository.java
@@ -3,5 +3,5 @@ package id.kings.stbmanager.repository;
 import id.kings.stbmanager.model.Apps;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppsRepository extends JpaRepository<Apps, Long> {
+public interface AppsRepository extends JpaRepository<Apps, Integer> {
 }
